Faster Abc settings
The settings dialog appeared slowly because of the voices tab. The voices tab is now split up into two stages: placing controls and filling controls. That last stage is only executed when selecting the voices tab. What makes it slow is filling the lookup list of comboboxes with the instrument names.