Fix determining gs_path on Linux. Fixes #66
When running EasyABC for the first time under Linux, it tries to
determine the path to gs from the output of `which gs`. However, in
python3, it returns a bytestring. EasyABC is expecting a str.  Simply
adding .decode() fixes the issue. The unicode() call is a remnant from
python2 and can be removed.